INFUND-713 replaced all Double with BigDecimal and fixed depending methods

+++ ifs-web-service/src/main/java/com/worth/ifs/application/finance/CostCategory.java
@@ -3,6 +3,7 @@ package com.worth.ifs.application.finance;
 import com.worth.ifs.application.finance.cost.CostItem;
 
 import javax.servlet.http.HttpServletRequest;
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -12,6 +13,6 @@ import java.util.List;
 public interface CostCategory {
     public List<CostItem> getCosts();
 
-    public Double getTotal();
+    public BigDecimal getTotal();
     public void addCost(CostItem costItem);
 }
